Earlier, a bug related to multi-threading was fixed in the UTF8DataChunk. A temp buffer was inadvertantly made static with no associated synchronized access to the variable which created the possibility of multiple thread each writing to the temp buffer at the same time causing data corruption.
The solution to this previous problem was to remove the
static (which wasn't meant to be there in the first place).
However, this meant that each data chunk contained its own
16K temp buffer which added a lot of extra memory used
during parsing of files.

Upon further investigation, it was discovered that the temp
buffer was only used within one method so the variable was
returned to static and access to the method was synchronized
to provide multi-thread safety.

No noticeable performance penalties arise from the use of
synchronized so, for now, this seems like a good solution
to the memory problem.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@315164 13f79535-47bb-0310-9956-ffa450edef68