Relaxed restraint of xml:space attribute for XHTML DTDs. The following is taken from the XML 1.0 Specification:
  [2.10] A special attribute named xml:space may be attached
  to an element to signal an intention that in that element,
  white space should be preserved by applications. In valid
  documents, this attribute, like any other, must be declared
  if it is used. When declared, it must be given as an
  enumerated type whose only possible values are "default"
  and "preserve". For example:

    <!ATTLIST poem   xml:space (default|preserve) 'preserve'>

Depending on how you read this text, you could argue that
when xml:space is declared, it *must* have those two options.
Or, a less relaxed reading would state that the only allowed
values are "default" and "preserve" but not necessarally
both.

Since the XHTML 1.0 Specification uses the relaxed reading
when defining xml:space attributes on elements, the
XMLDTDScanner has been modified to allow for the less
strict reading of the specification.

Thanks to David Brownell <david-b@pacbell.net> for the bug
report.

PR: 33


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@315344 13f79535-47bb-0310-9956-ffa450edef68