Phase one of implementing the <any/> element in XML Schema.
The content spec node was updated to include special ANY types. Then
the TraverseSchema was modified to build these content spec types
into the grammar when "compiling" the Schema. Then the grammar was
modified to be able to recognize the ANY types in the content spec
tree. This would shortcut the model into using a DFAContentModel and
a new CMNode tree was built. This content model node tree contained
a new kind of node, CMAll. Finally, the DFAContentModel was modified
in order to perform validation with these new ANY types.

As I said, this is phase one. We still have the problem with the
validator not knowing that something was matched as an ANY and
therefore not trying to resolve the element and validate it. In
addition, the "processContents" attribute on the <any/> element
is not implemented. Both of these problems will cause changes
to be made in the XMLValidator code. Fun fun fun!


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@315785 13f79535-47bb-0310-9956-ffa450edef68