I made a mistake in my previous comit.  There are Nodes which don't have the SYNCCHILDREN set to off.
When we deserialize a DOM we should never encounter such a case.

I suspect that the synchronizeChildren method is not doing what is suppose
to do which is to turn SYNCCHILDREN off once the nodes have been fluff up. This
is a bug and we should do further investigation.

This fix prevents from  this situation ever happening by  turning the SYNCCHILDREN flag off
as we desearialize the nodes in the readObject method.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@315886 13f79535-47bb-0310-9956-ffa450edef68