This fixes a problem with parameterizes entities such as: as the one found when we have in a dtd such thing as:
<!ENTITY % FOOBAR
           "'foofoo'" >
<!NOTATION test SYSTEM "">
<!ENTITY bar SYSTEM %FOOBAR; NDATA test >
<!ELEMENT FOO  (BAR?)>
<!ELEMENT BAR (MOCHA*)>
<!ELEMENT FOOFOO (#PCDATA)>

This bug also cause the parsing of a Schema containing a reference
to the structures or datatype dtd to fail with the following
message:
org.xml.sax.SAXException: Stopping after fatal error: The system identifier
must begin with either a single or double quote character.

A Caveat is that the XMLDTDScanner code is very obscure and we may have more
surprises.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@315895 13f79535-47bb-0310-9956-ffa450edef68