It is a well-formedness error to reference an undeclared parameter entity in a standalone document. Added check in XMLEntityManager#startEntity(String) and also added a method to allow the document scanner to set the standalone flag once it has parsed the XMLDecl line.
I also removed dead code in entity manager that was only
needed previously because the error handling was in the
document scanner. Now that it has been moved to the
startEntity method in the entity manager, those methods
for querying information about entities is not needed.
Perhaps later we'll discover a use for them and re-add
the methods but for now we'll keep the code clean by not
having them.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xerces_j_2@316334 13f79535-47bb-0310-9956-ffa450edef68