Applied patch from Lynn Monson:
I have found a significant bug in DOM Range creation.
The bug lies in RangeImpl.java in the checkIndex(..) method, which was doing
an improper test.  The bug makes it impossible to create a range with a
boundary point that lies within a text node. The problem can be seen with the
following code sample:

  DocumentImpl doc=new org.apache.xerces.dom.DocumentImpl();

  Element body = doc.createElement("body");
  doc.appendChild(body);
  Text text = doc.createTextNode("abcdef");
  body.appendChild( text );

  Range range = doc.createRange();


Patch reviewed by Elena Litani.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@316387 13f79535-47bb-0310-9956-ffa450edef68