Applied patch from Lynn Monson:
This brings the Xerces code into conformance
with the specification.  I've also included below, an example from previous
e-mail that demonstrates the bug that exists in the current code base.

take a very simple example:

Text node: abc
Let aRange span the character "a"
let sourceRange span the character "c"

If I invoke aRange.compareBoundaryPoints( START_TO_START, sourceRange ):

Xerces code says:

 if (how == START_TO_START) {
      endPointA = sourceRange.getStartContainer();
       endPointB = fStartContainer;
       offsetA = sourceRange.getStartOffset();
        offsetB = fStartOffset;
 } else

So far, so good.  "A" is sourceRange and "B" is aRange.  They have the same
container, and offsetA==2 and offsetB==0.  The next lines executed are:

   // case 1: same container
   if (endPointA == endPointB) {
       if (offsetA < offsetB) return -1;
       if (offsetA == offsetB) return 0;
       return 1;
   }

Since offsetA > offsetB, a +1 is returned.  But this contradicts the (new)
spec wording which says the result is:

  -1, 0 or 1 depending on whether the corresponding boundary-point of the
  Range is, respectively, before, equal to, or after the corresponding
  boundary-point of sourceRange

In the example, offsetB is the "corresponding boundary-point of the Range"
and offsetA is the "corresponding boundary-point of sourceRange."  Since
offsetB is before A, shouldn't the answer be -1 instead of +1?


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@316618 13f79535-47bb-0310-9956-ffa450edef68