Applied patch from Mark Diekhans:
This patch provides for hiding the NodeList DOM methods when necessary.
The approach adds a method getChildNodesUnoptimized() to ParentNode
that return an anonymous class implementing NodeList.  If a subclass
has a conflicting method getLength() or item(int); it must implement
getChildNodes() and then call this method to get a NodeList.
This was done for a couple of reasons:
   (1) Avoided duplicate code (in different files).
   (2) Avoided needing to make various private fields in ParentNode
       protected.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@316801 13f79535-47bb-0310-9956-ffa450edef68