Moved the "pipeline" setup from XMLDocumentScanner to the lowest classes in the hierarchy so that it can be changed without having to rewrite the whole SAX and DOM parser classes.
This is achieved through the introduction of two new classes
AbstractSAXParser and AbstractDOMParser which implement the appropriate
callbacks independently of the actual pipeline being used.

While this allows some code to be reused it is still not 100% satisfactory
because the pipeline setup code can no longer be shared between different
parsers. So we end up duplicating some code anyway. For this reason, We
might want to move to using some kind of "ParserConfiguration" object.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xerces_j_2@316900 13f79535-47bb-0310-9956-ffa450edef68