Added identity constraint code, ported from the v1 package.  Several things to note:
1.  Broke out an xpath package from the original "identity" package.
I did this so that, should we need the xpath code for something other
than identity constraint support (such as DOM level 3), it's there.

2.  I've merged the unique and key classes and also their traversers.
I did this because, from a schema compilation prospective, they're
virtually identical; the differences with respect to validation can be
accomodated with the type field of the base class.

3.  I've also changed the storage implementation in the grammar class
from being Vector'based to being array based.  This allows <keyref>s
to store indeces of the keys they reference rather than object
references.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@317576 13f79535-47bb-0310-9956-ffa450edef68