Introducing local namespace binder that only stores declarations for current namespace. This will allow to solve the following problem:
<root xmlns:ns1="http://example" xmlns="http://validdefault">
<foo:elem xmlns:ns1="newURI" xmlns="default"/>
</root>
If we don't remember what declarations we've seen on the local element we
won't know whether the declaration is a duplicate or it is redeclares previously
defined prefix.

We also store in the global namespace binder, the binding of an empty string to
an empty string (default decl xmlns="") to solve the following problem:

<elem xmlns="" xmlns="default"/>
Let's say a user modified the tree by adding xmlns="default" declaration.
We need to bind default namespace to some string (ei. empty string), to be
able to omit duplicate declarations for the same element (related to the problem
described above).

This also fixes a bug: http://nagoya.apache.org/bugzilla/show_bug.cgi?id=6516


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@318232 13f79535-47bb-0310-9956-ffa450edef68