Fixing Bug #20304: http://nagoya.apache.org/bugzilla/show_bug.cgi?id=20304
Static compile time and run time dependencies on
sun.io.CharToByteConverter made it impossible to compile
Xerces with JDKs which do not contain this class, and made
the serializer unusable (assuming you already have a binary)
for encodings which tried to access it. Users were probably
getting a NoClassDefFoundError in this situation.

For now, access the CharToByteConverter through reflection.
This should hopefully resolve both of those problems, however
on platforms where the converter is available we fall back to the
old behaviour of printing character references which are
expressable in the output encoding. This is something to revisit
in the future.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319453 13f79535-47bb-0310-9956-ffa450edef68