Fixing a bug which seems to have arisen from factoring entity scanner out from the entity manager several releases ago.
The buffer size passed to the entity scanner 
was not being kept in synch with the actual size
of the buffers in each scanned entity. If any
of the buffers were actually resized, it was possible
that the parser would throw an ArrayIndexOutOfBoundsException
for documents which contained names which are longer than
the current buffer size.

Conceivably the buffer size passed to entity scanner could be
used to determine a minimum size for resizing, if doubling the
buffer's size is smaller than this minimum. We currently don't
do this.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319455 13f79535-47bb-0310-9956-ffa450edef68