Re-implement XML11Configuration:
1. Remove dependency on the configurations that were designed to handle only XML 1.0 documents.
Relying on those configurations might have caused several bugs in reset, configurePipeline.
2. Instead of having one vector of components, user 3 vectors: one for XML 1.0 components, one for
XML 1.1. components, and one vector for the components that do not depend on XML version, such as XMLEntityManager
3. Add a new internal feature "internal/parser-settings" that is set by the configuration to notify the components if anything in the configuration settings have been changed.

Make sure that components in the pipeline query "internal/parser-settings" feature to find out if they need to reset features and properties.
This change improves the performance for non-validating SAX parser by ~14%. It improves schema validation with grammar cashing by ~9%.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319463 13f79535-47bb-0310-9956-ffa450edef68