Redistributing code for content negotation.
We were reading the URLConnection to early.  We cannot set 
those parameters in XIncludeHandler because we may lose some
information if we process the URLConnection and then open a stream.  

The parser needs a URLConnection to follow HTTP redirects and may 
also need this object in the future to read external 
encoding information which may be available in the Content-Type.
We may also need to read the URLConnection to determine
the encoding of a text include.

Also fixed a bug involving the case of the attributes on the include
element. Accept-Language -> accept-language.

REVISIT: If we're going to support content negotation for
parse=xml we need to find a clean way to propogate the
parameters to the child parser.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319666 13f79535-47bb-0310-9956-ffa450edef68