Fixing Bug #24579: http://nagoya.apache.org/bugzilla/show_bug.cgi?id=24579
XML 1.0 SE - E27. According to Unicode 3.1 conformant UTF-8
interpreters must reject non-shortest form byte sequences.
Some examples are C0 80 and E0 80 80, bothl corresponding
to codepoint 0.

Extra checks are required for all multi-byte sequences. Currently
this is done with anding the bytes with a mask. We should revisit
this code to check whether it would be faster to check the value
of the character after the bytes have been combined.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319679 13f79535-47bb-0310-9956-ffa450edef68