Improve performance of attribute value normalization for XML 1.0 documents.
Check if each character is < 0x20 instead of calling XMLChar.isSpace.
This simple check against 0x20 is safe because at the point we 
normalize the attribute value we've determined that it is well formed, 
and therefore if we're looking at anything less then 0x20 it must be 
TAB, LF or CR. This would avoid a method call and table lookup and 
allow us to skip over 0x20 which doesn't require further normalization.

Since this doesn't work for XML 1.1 docs since control chars are
allowed in them add the old method to the 1.1 scanners.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319700 13f79535-47bb-0310-9956-ffa450edef68