Improve performance of the initialization of the character flag arrays for XMLChar and XML11Char. Although these arrays are always the same we were building them dynamically.
Created some new utility classes for generating a
more efficient initialization of these arrays. These
generators should be used again if XMLChar or
XML11Char need to be updated. The generated
code uses java.util.Arrays (from JDK 1.2) to
fill chunks of the array with the same value. This
is similar to the C library's memset.

This should significantly improve class initialization
for XMLChar and XML11Char which is important
for applications who care about faster startup and
it is also important if these classes are garbage 
collected and reloaded several times.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319746 13f79535-47bb-0310-9956-ffa450edef68