Fixing Bug #27081: http://nagoya.apache.org/bugzilla/show_bug.cgi?id=27081
Fixing a reset bug.

When the parser scans sequences of ']'
in character data it appends them to a buffer before reporting
them in case the section contains an illegal ']]>'. While in this
section a boolean is set indicating that content is being
scanned in case the current entity ends. This makes it possible
to flush the buffer to a characters call before calling endEntity.

If an exception occured in this section, this state was not being
reset on a new parse. The buffer to scan content is also used
to scan comments and a few other structures. While in this
broken state the parser would report anything currently in the
buffer as characters at the end of any general entity as well
as the document entity.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319802 13f79535-47bb-0310-9956-ffa450edef68