Fixing Bug #26685: http://nagoya.apache.org/bugzilla/show_bug.cgi?id=26685
SAX requires that when an entity is skipped that it be reported
to ContentHandler.skippedEntity.

From the API spec for ContentHandler:
"The Parser will invoke this method each time the entity is skipped. 
Non-validating processors may skip entities if they have not seen 
the declarations (because, for example, the entity was declared in 
an external DTD subset). All processors may skip external entities, 
depending on the values of the 
http://xml.org/sax/features/external-general-entities and
the http://xml.org/sax/features/external-parameter-entities 
properties."

XNI (which isn't going to change) has no built-in method for 
communicating skipped entities. XMLEntityManager needed
some method to communicate that an entity has been
skipped so I've added Augmentations to our internal
interface XMLEntityHandler. This is consistent with
interfaces in XNI.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319821 13f79535-47bb-0310-9956-ffa450edef68