Fixing Bug #27083 & #27422: http://nagoya.apache.org/bugzilla/show_bug.cgi?id=27083 http://nagoya.apache.org/bugzilla/show_bug.cgi?id=27422
Malformed UTF-8 and US-ASCII byte sequences encountered
while parsing a document are well-formedness errors. Report
them to the error reporter like every other well-formedness
violation instead of letting the IOException propogate the
user.

This fixes a bug in XInclude processing where the processor
would treat an IOException thrown because one of our
readers detected a bad byte sequence as a ResourceError,
causing the XInclude processor to look for a fallback to
parse instead of failing. The result of this would be a
malformed stream of events.

For ASCII, we were only using our built-in reader if the 
encoding specified was the canonical name US-ASCII. For
aliases the Java InputStreamReader was being picked up
which doesn't seem to report bad ASCII bytes. Now we
look into the EncodingMap so that we cover all of ASCII's
aliases.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319832 13f79535-47bb-0310-9956-ffa450edef68