Fixing Jira Bug #948: http://nagoya.apache.org/jira/browse/XERCESJ-948
In DOMNormailzer and DOMSerializerImpl DOMLocator.relatedNode
and DOMError.relatedData were not being reported correctly in
many instances. Patched thanks to Naela Nissar.

Also fixed a bug in DOMSerializerImpl where the well-formedness
of an element name was checked over and over again instead of
checking the well-formedness of the names of its attributes.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319894 13f79535-47bb-0310-9956-ffa450edef68