Fixing Jira Bug #960: http://nagoya.apache.org/jira/browse/XERCESJ-960
Two new kinds of components (XSFacet & XSMultiValueFacet)
were added to the XML Schema API some time ago, but the 
array in SchemaGrammar was't updated to reflect such change.

Fixing another bug where we returned null where an empty
XSNamedMap is expected for getComponents and 
getComponentsByNamespace. This wasn't updated when
the spec changed.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319919 13f79535-47bb-0310-9956-ffa450edef68