Fixing Jira Bug #525: http://nagoya.apache.org/jira/browse/XERCESJ-525
Instances of some of the classes in the HTML DOM implementation
have a reference to a HTMLCollection. When these nodes are
cloned the reference to the old collection was being kept, so
methods which return a collection were referring to nodes in the
old subtree instead of the newly cloned subtree. This is fixed
thanks to the patch by Aaron Kardell.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319929 13f79535-47bb-0310-9956-ffa450edef68