Fixing a bug reported to xerces-j-user by Robert Houben: http://marc.theaimsgroup.com/?l=xerces-j-user&m=108587697229050&w=2
We were performing reference comparison with the namespace name
passed into getComponentsByNamespace. The XML Schema API does not
constrict users to using internalized strings so we must use equals.

Should also be returning an empty map when a grammar does not exist
for a given namespace instead of null.

Both of these problems are now fixed.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319932 13f79535-47bb-0310-9956-ffa450edef68