Grammar Caching and SAX's EntityResolver2:
Do not bother checking the grammar pool if no public or system identifier was provided. 
Since so many different DTDs have roots in common, using only a root name as the 
key may cause the wrong grammar to be retrieved from the grammar pool. This scenario
would occur when an ExternalSubsetResolver has been queried and the
XMLInputSource returned contains an input stream but no external identifier.
This can never happen when the instance document specified a DOCTYPE.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319934 13f79535-47bb-0310-9956-ffa450edef68