Fixing Jira Bug #470: http://nagoya.apache.org/jira/browse/XERCESJ-470
The call to a private super constructor in BuiltinSchemaGrammar
seems to violate VM specifications. An anonymous class:
SchemaGrammar$1.class seems to be related to a private
constructor declared on SchemaGrammar which is implicitly called
from a constructor in the static nested class BuiltinSchemaGrammar
(which extends SchemaGrammar). This class appears to be
malformed [1].

Changing the access modifier from private to protected on the
no arguments constructor of SchemaGrammar to eliminate this
anonymous class.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=67600


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319952 13f79535-47bb-0310-9956-ffa450edef68