Fixing Jira Bug #843: http://nagoya.apache.org/jira/browse/XERCESJ-843
To ensure that only the aggregate document is schema validated, the 
XInclude handler sets the schema feature to false to keep the child 
pipeline it creates from containing a schema validator. This will activate 
the DTD validator on the child pipeline if the validation feature was set 
to true, causing the include to be validate against a (probably
non-existent) DTD.

This should now be fixed. In addition to turning the schema validation
feature off, turn off other related validation features.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@319991 13f79535-47bb-0310-9956-ffa450edef68