JIRA Issue #998: http://nagoya.apache.org/jira/browse/XERCESJ-998
Improve the performance of the symbol table by
dynamically resizing it and rehashing it once it
reaches some threshold. Previously the SymbolTable
did not scale well, since the size of the table was
fixed.  Neeraj had started looking at this a couple
years ago.  Thanks to John Kim for contributing
this performance improvement.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320003 13f79535-47bb-0310-9956-ffa450edef68