JIRA Issue #1001: http://nagoya.apache.org/jira/browse/XERCESJ-1001
When processing attributes in the scanner we store both the
non-normalized and normalized values of an attribute. Attribute
value normalization as specified by the XML 1.0 specification
replaces tabs and line breaks with spaces and performs entity
expansion, if any entity references exist in the attribute values.

In the vast number of real world documents both the
non-normalized and normalized values are the same, however we
were creating a new string for both values. Thanks to the
patch by John Kim, we only create two strings if the non-normalized
and normalized value are different.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320019 13f79535-47bb-0310-9956-ffa450edef68