Fixing JIRA Bug #816: http://nagoya.apache.org/jira/browse/XERCESJ-816
This had only been partially fixed. A value for the encoding
attribute specified on an LSInput is authoritative, meaning
it will override any encoding in an XML or Text declaration.
Should be covering all of these cases now. Renamed some
methods in XMLEntityManager.ScannedEntity so that their
names are more descriptive of what they actually do.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320062 13f79535-47bb-0310-9956-ffa450edef68