Fixing JIRA Bug #1039: http://issues.apache.org/jira/browse/XERCESJ-1039
To validate an anyURI we apply the escaping rules from XLink to make it a URI
that we then validate with Xerces' URI class.  We were returning the escaped
value instead of the original as the actual value. i.e. "http://www.exa mple.com"
was being returned as "http://www.exa%20mple.com".


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320210 13f79535-47bb-0310-9956-ffa450edef68