Fixing JIRA Bug #1044: http://issues.apache.org/jira/browse/XERCESJ-1044
Namespace lookup was incorrectly using the index into the context array rather
than the member of that array, causing the processor to report that mappings
don't exist when in fact they do exist. This bug is fixed thanks to the patch
by Eric Isakson. There were three other methods where this was also a
problem.  I've fixed those as well.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320226 13f79535-47bb-0310-9956-ffa450edef68