DOM validation support.
In the RI this was achieved using identity transformers.
Here we go directly from DOM to XNI and back
instead of the intermediate SAX step. In addition to eliminating
the dependency on a transformer this gives us better performance.
We do a non-recursive walk of the DOM to prevent a stack overflow
for DOM's with high depth.

It should also be noted that the result builder will fill in PSVI and
DOM TypeInfo if the result node comes from the Xerces' DOM
implementation.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320386 13f79535-47bb-0310-9956-ffa450edef68