Completing the implementation of reset.
Created an extension of org.apache.xerces.parsers.SAXParser
which tracks changes to features and properties. This allows
us to reset the parser to its initial state without having to
dig deep into the parser configuration which in general isn't
possible. On reset we only restore the features and properties
which changed since the last reset.

Also added support for setting the secure processing
feature directly on the XMLReader.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320488 13f79535-47bb-0310-9956-ffa450edef68