Fixing JIRA Issue #1090: http://issues.apache.org/jira/browse/XERCESJ-1090
Clause 2.1 of Section 3.4.4 of Part 1: Structures says:  "If the {content type} is empty,
then the element information item has no character or element information item [children]".

Section 3.1.4 says: "... comments and processing instructions, even in the
midst of text, are ignored for all �validation� purposes".

The schema validator was reporting an error if a comment or processing instruction was
present as a child of an element declared to have empty content. This behaviour
was introduced in CVS v1.28 when it was thought that errors should be reported for
such content.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320551 13f79535-47bb-0310-9956-ffa450edef68