Fixing JIRA Issue #1089: http://issues.apache.org/jira/browse/XERCESJ-1089
When both the schema validation and validation features have been
turned on, to be consistent with what occurs on the main pipeline,
DTD validation should occur if the document being processed on the
child pipeline has a DOCTYPE. If no DOCTYPE is present then no
errors should be reported. This is controlled by turning on the
dynamic validation feature.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320556 13f79535-47bb-0310-9956-ffa450edef68