Attempting to fix JIRA Issues #1054 & #1076: http://issues.apache.org/jira/browse/XERCESJ-1054 http://issues.apache.org/jira/browse/XERCESJ-1076
The first time match() is called on a RangeToken, a map is created for
fast lookups. There was a race condition which made it possible for
the map to be refill while another thread was reading from it. Though
I've never been able to reproduce either JIRA issue, it looks like
this could have been the problem. Now we assign the map only
once it's been filled.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@320567 13f79535-47bb-0310-9956-ffa450edef68