Eliminating the separate DTD validator. It was being used even with validation off for any DOM containing a DocumentType node which is just a waste of cycles. Simply having a DocumentType node in a DOM was leading to the DTD being processed every time Document.normalizeDocument() is called even if DTD validation isn't turned on. This could be extremely slow especially if the system id in the DOCTYPE points to a resource somewhere on the internet.
Now we treat DTD validation like schema validation, that is we retrieve a validation 
handler from the DOMImplementation, setup up the validation handler and then invoke 
the XMLDocumentHandler on the validation handler as we walk the DOM tree. We now also 
call comment() and processingInstruction() on the validation handler so that a DTD 
validatior can determine whether a comment or PI appears in EMPTY content.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@349818 13f79535-47bb-0310-9956-ffa450edef68