Fixing a bug. The DOM parser sets the input encoding on the document during the startDocument event. The problem is that this occurs before the XML declaration has been read so the encoding stored in the DOM may be the auto-detected encoding and not the actual encoding of the document.
For a document such as:
<?xml version="1.0" encoding="US-ASCII"?>
<root/>

the actual encoding was being reported as UTF-8 instead of US-ASCII.

The actual encoding of the document may not be determined by the scanner until after the xmlDecl 
event has been reported down the pipeline. To get the "final" encoding we now read the locator in 
endDocument() and set the encoding it returns on the document.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@377247 13f79535-47bb-0310-9956-ffa450edef68