JIRA Issue #1143: http://issues.apache.org/jira/browse/XERCESJ-1143
JAXP 1.4 clarifies that features set on the SchemaFactory should be propogated
to the Schemas created from the factory and in turn ValidatorHandlers and
Validators constructed from those Schema objects. Feature propogation is now
supported thanks to the patch by Peter McCracken which I've modified slightly
so that the secure processing feature is set correctly.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@381134 13f79535-47bb-0310-9956-ffa450edef68