Fixing JIRA Bug #1007: http://issues.apache.org/jira/browse/XERCESJ-1007
When serializing a DOM which only implements DOM Level 2 calling getXmlVersion(), etc...
by reflection will cause a AbstractMethodError to be thrown. The Error was not being
caught. This is fixed thanks to the patch by Nathan Beyer. I've enhanced it a bit
so that the methods are only looked up once and made sure that ThreadDeath is 
re-thrown if it's caught.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@381247 13f79535-47bb-0310-9956-ffa450edef68