Fixing JIRA Bug #1137: http://issues.apache.org/jira/browse/XERCESJ-1137
The code which corrects field XPaths so that they start with "./" did not
handle a union of XPath expressions. Only the first XPath in the union was 
being fixed up. I've fixed that problem though the code should probably
be replaced with a more efficient solution since it could generate many
temporary objects.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@381279 13f79535-47bb-0310-9956-ffa450edef68