Fixing JIRA Bug #1180: http://issues.apache.org/jira/browse/XERCESJ-1180
When an element node is materialized we were assuming that it must be the
document element if (docElement == null). It's possible for this field to
be null when other elements are created, for instance if the children of the
document element are expanded after the document element has been removed
from the tree. In such a case one of the child element nodes was being
blindly assigned as the document node. This should be fixed now. Note that
DeferredDocumentImpl.synchronizeChildren() handles assignment of the
document element so the code that I'm removing was always redundant.


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@437729 13f79535-47bb-0310-9956-ffa450edef68