Fixing JIRA Issue #1182: http://issues.apache.org/jira/browse/XERCESJ-1182
There was a race condition which allowed the XMLStreamReader constructor to
return before the start document event is produced. I've fixed that as well
as significantly simplifying the thread handling. There was a boolean that
was used to track which thread was running. Since only one thread is running
at a time we don't need it.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/stax-dev@451674 13f79535-47bb-0310-9956-ffa450edef68