Fixing JIRA Bug #1239: http://issues.apache.org/jira/browse/XERCESJ-1239
If the union of two attribute wildcards is not expressible an error should be reported.
Note that XML Schema 1.0 2nd edition doesn't actually specify this constraint. It's a bug [1]
in the spec which will eventually be fixed. At the moment we're just guessing what the error 
code will be. If it turns out to be something else we'll need to change it.

[1] http://www.w3.org/Bugs/Public/show_bug.cgi?id=4390

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@518306 13f79535-47bb-0310-9956-ffa450edef68