Fixing JIRA Bug #1079 and #1237: http://issues.apache.org/jira/browse/XERCESJ-1079 http://issues.apache.org/jira/browse/XERCESJ-1237
If a type derived from xs:integer specifies a fractionDigits facet with a value other than one
its an error. We weren't reporting one before.

"0" was always being returned as the lexical value for the totalDigits facet of a type derived
from xs:integer. This should be the value returned for fractionDigits.

If a fractionDigits facet is specified on a type derived from xs:integer we need to take special
care that we only create one instance of XSFacet for fractionDigits.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@518311 13f79535-47bb-0310-9956-ffa450edef68