JIRA Issue #1264: http://issues.apache.org/jira/browse/XERCESJ-1264
As part of its normal control flow the XMLEntityScanner will throw an EOFException when it reaches the end of the document. For small documents, this can take up as much as 20-25% of the total execution time in the parser but could be even more than that depending on the depth of the call stack at the point where parse() is called. To eliminate most of this cost we now cache the exception and avoid filling in its stack trace (thus also saving some memory).


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@568405 13f79535-47bb-0310-9956-ffa450edef68