Fixing JIRA Issue #1296: http://issues.apache.org/jira/browse/XERCESJ-1296
It was possible for getBaseURI() called on namespace-aware Element nodes
to have an exponential execution time relative to the depth of the node.
This would occur when the Document node had no URI or one of the ancestors
had a a base URI which isn't valid. Thanks to Ludger Bünger for pointing
out this issue.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xml-schema-1.1-dev@645346 13f79535-47bb-0310-9956-ffa450edef68