JIRA Issue #1298: http://issues.apache.org/jira/browse/XERCESJ-1298
Store user data in a WeakHashMap to allow it and the node associated with it
to be garbage collected if no other references exist for the node. Thanks to
Ludger Bünger for the patch. I made a few modifications and added readObject(),
writeObject() methods to preserve object serialization compatibility with
previous versions of Xerces.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@695558 13f79535-47bb-0310-9956-ffa450edef68