Fixing JIRA Issue #1334: http://issues.apache.org/jira/browse/XERCESJ-1334
If the name of element in the document below the root element happens to be match the
name of the root element we were not passing this node to the acceptNode() method of
an LSParserFilter. This was occurring because we were doing the check based on the name
of the node instead of identity with the root element object. This is fixed now thanks
to the patch from Arthur De Magalhaes.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xml-schema-1.1-dev@709531 13f79535-47bb-0310-9956-ffa450edef68