doing a little improvement (thought this was useful, and is optional) to assertions error messages (which uses Xerces extensions namespace). we are now able to provide an atomic value available in validation context (the value of assertions XPath2 context variable $value) to assertions error messages.
for example, if there is an assertion facet defined like following:
<xs:assertion test="$value mod 2 = 0" xerces:message="The value must be divisible by 2. Atomic value {$value} is not divisible by 2." />

and if this assertion fails in an validation episode, the following error message is generated by Xerces:
"The value must be divisible by 2. Atomic value 101 is not divisible by 2."
(the instance document had value 101)

The placeholder for this feature must be {$value} and there can be multiple instances of this placeholder in single message, and all of them will be replace by the run-time value of $value. 


git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xml-schema-1.1-dev@1029000 13f79535-47bb-0310-9956-ffa450edef68