Fixes to assertions evaluations on elements with simpleType's specifying default values.
It seems that the following element declaration,

<xs:element name="x">
  <xs:complexType>
     <xs:sequence>
        <xs:element name="y" type="xs:string" default="hi" />
     </xs:sequence>
     <xs:assert test="y = 'hi'" />
  </xs:complexType>
</xs:element>

should successfully validate the following instance fragment,

<x>
  <y/>
</x>

(I believe that XML schema assertions should have access to the default value of the element, in case the instance document doesn't provide a value for such an element)

Prior to this fix, a text node that needs to be constructed from the default value specified in the schema (with value "hi" in this case) wasn't part of the XDM tree on which assertions evaluated.

This fix solves this issue.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xml-schema-1.1-dev@1056276 13f79535-47bb-0310-9956-ffa450edef68