committing few improvements for inheritable attributes processing.
It seems we should probably have something like following data structures,

XMLAttributes[], or
XMLSchemaValidator -> InheritableAttribute[]

in the field PSVIElementNSImpl -> fInheritedAttributes instead of the present XSAttributeUse[] (which only provides the attribute schema component and not the attribute value).

This would allow both the attribute QName and it's value for inherited attributes to be available in PSVI results after validation episode.

I might attempt above proposed improvement in near future.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xml-schema-1.1-dev@1086899 13f79535-47bb-0310-9956-ffa450edef68