committing few improvements for inheritable attributes PSVI results.
creating a new class (XMLAttribute) to store attribute declaration and it's value to produce information for inherited attributes PSVI results. i did not use XMLAttributes class for this purpose, since it has lot of information that's unnecessary for this purpose, and may require more lines-of-code by applications using the PSVI results.

The inner class InheritableAttribute in XMLSchemaValidator is not needed as a consequence of this change and is now deleted.

git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xml-schema-1.1-dev@1087961 13f79535-47bb-0310-9956-ffa450edef68